Correct test placement
